Priority tie-break between multiple peerings to the same node
This allows specifying a `prio` number (lower is better) which is used to prioritise certain links when multiple peerings to the same node exist. It does not prioritise peerings between different nodes in any way. If the multiple links do not have distinct priorities (i.e. they are all zero or otherwise equal) then this code has no side effects.

It works by tie-breaking the tree routing using the priority field after the other checks, so that if we come across another higher priority link to the same next-hop node, we'll use it instead. In theory this should influence the path setup packets in particular, pushing them over higher priority links, resulting in the paths themselves being set up over those higher priority links.

It also adds another loop into the DHT routing to check that a higher priority link for the next-hop key is available, so that bootstrap, bootstrap ACK etc packets will also be prioritised.

In theory we could allow users to specify their own priorities in the peering URI or at the multicast scope level or similar, or we could do so automatically in some cases by link type (multicast vs statically configured). It could also be used to specifically deprioritise slower link types in the future (Bluetooth, for example).

This specifically does not modify the parent selection because it is quite likely that doing so would result in more spanning tree traffic in cases where the priority link isn't the lowest latency path to the root. If that were the case, you'd switch parent on a new tree update to the lowest latency link, and then you'd do so *again* when the higher priority link receives the update afterwards. It isn't really clear that the parent relationship really matters if we influence the forwarding pattern anyway.